[SG] u16 & i16

+++ src/algo/heuristics.rs
@@ -14,7 +14,7 @@ impl fmt::Display for Heuristic {
     }
 }
 
-pub fn heuristic(heuristic: &Heuristic, size: i8, state: &Vec<i8>, target: &Vec<i8>) -> u32 {
+pub fn heuristic(heuristic: &Heuristic, size: u16, state: &Vec<u16>, target: &Vec<u16>) -> u32 {
    match heuristic {
 	  Heuristic::Manhattan => {
 		return manhattan(size, state, target);
@@ -31,7 +31,7 @@ pub fn heuristic(heuristic: &Heuristic, size: i8, state: &Vec<i8>, target: &Vec<
    }
 }
 
-fn manhattan(size: i8, state: &Vec<i8>, target: &Vec<i8>) -> u32 {
+fn manhattan(size: u16, state: &Vec<u16>, target: &Vec<u16>) -> u32 {
     let mut dist: u32 = 0;
     for i in 0..(state.len()){
         if state[i] != size * size {
@@ -44,7 +44,7 @@ fn manhattan(size: i8, state: &Vec<i8>, target: &Vec<i8>) -> u32 {
     return dist;
 }
 
-fn euclidian(size: i8, state: &Vec<i8>, target: &Vec<i8>) -> u32 {
+fn euclidian(size: u16, state: &Vec<u16>, target: &Vec<u16>) -> u32 {
     let mut dist: f32 = 0.0;
     for i in 0..(state.len()) {
         if state[i] != size * size {
@@ -57,7 +57,7 @@ fn euclidian(size: i8, state: &Vec<i8>, target: &Vec<i8>) -> u32 {
     return dist as u32;
 }
 
-fn hamming_distance(size: i8, state: &Vec<i8>, target: &Vec<i8>) -> u32 {
+fn hamming_distance(size: u16, state: &Vec<u16>, target: &Vec<u16>) -> u32 {
     let mut misplaced: u32 = 0;
     for i in 0..(size * size) {
         if state[i as usize] != size * size {
@@ -69,17 +69,17 @@ fn hamming_distance(size: i8, state: &Vec<i8>, target: &Vec<i8>) -> u32 {
     return misplaced;
 }
 
-fn is_conflicting(tile_a: i8, target_a: i8, tile_b: i8, target_b: i8) -> bool {
+fn is_conflicting(tile_a: u16, target_a: u16, tile_b: u16, target_b: u16) -> bool {
     return (tile_b > tile_a && target_b < target_a) || (tile_a > tile_b && target_a < target_b);
 }
 
-fn find_conflicting_tiles_nb_for(size: i8, tile: usize, target: usize, line: &Vec<i8>, target_line: &Vec<i8>) -> i8 {
-    let mut conflicts_nb: i8 = 0;
+fn find_conflicting_tiles_nb_for(size: u16, tile: usize, target: usize, line: &Vec<u16>, target_line: &Vec<u16>) -> u16 {
+    let mut conflicts_nb: u16 = 0;
     if line[tile as usize] != size * size {
         for index in 0..size as usize {
             let target_b = target_line.iter().position(|&x| x == line[index as usize]);
             if target_b.is_some() {
-                if line[index] != size * size && is_conflicting(tile as i8, target as i8, index as i8, target_b.unwrap() as i8) {
+                if line[index] != size * size && is_conflicting(tile as u16, target as u16, index as u16, target_b.unwrap() as u16) {
                     conflicts_nb += 1;
                 }
             }
@@ -88,9 +88,9 @@ fn find_conflicting_tiles_nb_for(size: i8, tile: usize, target: usize, line: &Ve
     return conflicts_nb;
 }
 
-fn line_extra_moves(size: i8, line: &Vec<i8>, target_line: Vec<i8>) -> i8 {
-    let mut total_conflicting_tiles: i8 = 0;
-    let mut conflicts_table: Vec<i8> = Vec::new();
+fn line_extra_moves(size: u16, line: &Vec<u16>, target_line: Vec<u16>) -> u16 {
+    let mut total_conflicting_tiles: u16 = 0;
+    let mut conflicts_table: Vec<u16> = Vec::new();
     for tile_idx in 0..size as usize {
         let target = target_line.iter().position(|&x| x == line[tile_idx]);
         if target.is_some() {
@@ -99,14 +99,14 @@ fn line_extra_moves(size: i8, line: &Vec<i8>, target_line: Vec<i8>) -> i8 {
             conflicts_table.push(0);
         }
     }
-    while conflicts_table.iter().sum::<i8>() > 0 {
+    while conflicts_table.iter().sum::<u16>() > 0 {
         let most_conflicting_tile: usize = conflicts_table.iter().position(|&x| x == *conflicts_table.iter().max().unwrap_or(&0)).unwrap_or(0);
         conflicts_table[most_conflicting_tile] = 0;
         for tile in 0..size as usize {
             let target = target_line.iter().position(|&x| x == line[tile]);
             let target_b = target_line.iter().position(|&x| x == line[most_conflicting_tile]);
             if target.is_some() && target_b.is_some() {
-                if is_conflicting(tile as i8, target.unwrap_or(tile) as i8, most_conflicting_tile as i8, target_b.unwrap_or(most_conflicting_tile) as i8) {
+                if is_conflicting(tile as u16, target.unwrap_or(tile) as u16, most_conflicting_tile as u16, target_b.unwrap_or(most_conflicting_tile) as u16) {
                     conflicts_table[tile] -= 1;
                 }
             }
@@ -116,16 +116,16 @@ fn line_extra_moves(size: i8, line: &Vec<i8>, target_line: Vec<i8>) -> i8 {
     return total_conflicting_tiles;
 }
 
-fn linear_conflict(size: i8, state: &Vec<i8>, target: &Vec<i8>) -> u32 {
-    let mut extra_moves: i8 = 0;
+fn linear_conflict(size: u16, state: &Vec<u16>, target: &Vec<u16>) -> u32 {
+    let mut extra_moves: u16 = 0;
     for row_index in 0..size {
-        let row: Vec<i8> = Vec::from(&state[((size * row_index) as usize)..((size * row_index + size) as usize)]);
-        let target_row: Vec<i8> = Vec::from(&target[((size * row_index) as usize)..((size * row_index + size) as usize)]);
+        let row: Vec<u16> = Vec::from(&state[((size * row_index) as usize)..((size * row_index + size) as usize)]);
+        let target_row: Vec<u16> = Vec::from(&target[((size * row_index) as usize)..((size * row_index + size) as usize)]);
         extra_moves += line_extra_moves(size, &row, target_row);
     }
     for col_index in 0..size {
-        let col: Vec<i8> = state.iter().cloned().enumerate().filter(|&(i, _)| i % size as usize == col_index as usize).map(|(_, e)| e).collect();
-        let target_col: Vec<i8> = target.iter().cloned().enumerate().filter(|&(i, _)| i % size as usize == col_index as usize).map(|(_, e)| e).collect();
+        let col: Vec<u16> = state.iter().cloned().enumerate().filter(|&(i, _)| i % size as usize == col_index as usize).map(|(_, e)| e).collect();
+        let target_col: Vec<u16> = target.iter().cloned().enumerate().filter(|&(i, _)| i % size as usize == col_index as usize).map(|(_, e)| e).collect();
         extra_moves += line_extra_moves(size, &col, target_col);
     }
     return manhattan(size, state, target) + (2 * extra_moves) as u32;
